chore: move PathNodesBuilder and pathNode DSL to its own file chore: add close property to PathNodesBuilder enabling close command
