+++ shared/src/commonMain/kotlin/dev/tonholo/s2c/domain/PathNodes.kt
@@ -415,127 +415,3 @@ sealed class PathNodes(
         }
     }
 }
-
-@DslMarker
-annotation class PathNodesDsl
-
-@PathNodesDsl
-class PathNodesBuilder(val command: Char) {
-    private val commandMap = mapOf(
-        PathNodes.MoveTo.COMMAND to PathNodes.MoveTo.COMMAND_SIZE,
-        PathNodes.ArcTo.COMMAND to PathNodes.ArcTo.COMMAND_SIZE,
-        PathNodes.VerticalLineTo.COMMAND to PathNodes.VerticalLineTo.COMMAND_SIZE,
-        PathNodes.HorizontalLineTo.COMMAND to PathNodes.HorizontalLineTo.COMMAND_SIZE,
-        PathNodes.LineTo.COMMAND to PathNodes.LineTo.COMMAND_SIZE,
-        PathNodes.CurveTo.COMMAND to PathNodes.CurveTo.COMMAND_SIZE,
-        PathNodes.ReflectiveCurveTo.COMMAND to PathNodes.ReflectiveCurveTo.COMMAND_SIZE,
-        PathNodes.QuadTo.COMMAND to PathNodes.QuadTo.COMMAND_SIZE,
-        PathNodes.ReflectiveQuadTo.COMMAND to PathNodes.ReflectiveQuadTo.COMMAND_SIZE,
-    )
-
-    init {
-        check(command in commandMap.keys) {
-            "The command $command is not a Path command."
-        }
-    }
-
-    private var args: List<String>? = null
-    var isRelative: Boolean = false
-    var minified: Boolean = false
-
-    fun args(vararg args: Any) {
-        val parsedArgs = args.map { arg ->
-            when (arg) {
-                is String, is Number, is Char -> arg.toString()
-                is Boolean -> if (arg) "1" else "0"
-                else -> error("The argument type of ${arg::class} is unsupported.")
-            }
-        }
-        this.args = parsedArgs
-    }
-
-    private fun enforceArgumentSize(args: List<String>): List<String> {
-        val hasCloseCommand = args.last() == CLOSE_COMMAND
-        val size = if (hasCloseCommand) args.size - 1 else args.size
-        check(size == commandMap[command]) {
-            "Invalid number of arguments. Expected ${commandMap[command]}, given ${args.size}"
-        }
-
-        return if (hasCloseCommand) {
-            args.dropLast(1)
-                .mapIndexed { index: Int, s: String ->
-                    if (index == size - 1) "${s}z" else s
-                }
-        } else {
-            args
-        }
-    }
-
-    fun build(): PathNodes {
-        val args = checkNotNull(args) {
-            "Missing path arguments."
-        }
-
-        return when (command) {
-            PathNodes.MoveTo.COMMAND -> PathNodes.MoveTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.ArcTo.COMMAND -> PathNodes.ArcTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.VerticalLineTo.COMMAND -> PathNodes.VerticalLineTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.HorizontalLineTo.COMMAND -> PathNodes.HorizontalLineTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.LineTo.COMMAND -> PathNodes.LineTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.CurveTo.COMMAND -> PathNodes.CurveTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.ReflectiveCurveTo.COMMAND -> PathNodes.ReflectiveCurveTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.QuadTo.COMMAND -> PathNodes.QuadTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            PathNodes.ReflectiveQuadTo.COMMAND -> PathNodes.ReflectiveQuadTo(
-                values = enforceArgumentSize(args),
-                isRelative = isRelative,
-                minified = minified,
-            )
-
-            else -> error("Unsupported path command $command")
-        }
-    }
-}
-
-internal fun pathNode(command: PathCommand, configuration: PathNodesBuilder.() -> Unit): PathNodes =
-    PathNodesBuilder(command).apply(configuration).build()
-
